MapView/requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.20.1
idna==3.6
Kivy==2.3.0
kivy_garden.mapview==1.0.6
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
